Improve RecoveryPolicy to support more flexible use cases
[Breaking Change]

refs #35
